reorganize comps
