+++ src/App.tsx
@@ -9,7 +9,7 @@ import { Store } from "./store/Store";
 import Header from "./components/layout/Header";
 import HomePage from "./components/routes/HomePage";
 import SubcategoryList from "./components/SubcategoryList";
-import ProductList from "./components/ProductList";
+import ProductList from "./components/products/ProductList";
 
 import categories from "./content.json";
 import products from "./items.json";
