+++ src/components/products/ProductItem.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Card } from "react-bootstrap";
 
-import { Product } from "../store/Store";
+import { Product } from "../../store/Store";
 
 import styles from "./ProductItem.module.css";
 import ProductParameters from "./ProductParameters";
@@ -13,7 +13,7 @@ interface ProductItemProps {
 const ProductItem: React.FC<ProductItemProps> = ({ product }) => {
   return (
     <Card className={styles.productCard}>
-      <img src={product.img} />
+      <img src={product.img} alt="phone_img" />
       <div className={styles.desc}>
         <h3>{product.name}</h3>
         <ProductParameters product={product} />
